refactor:  EditOurWorkEntryCell and OurWorkEntriesTable components
