+++ src/app/_components/PortalComponents/EditOurWorkEntryCell/DeleteOurWorkEntry.tsx
@@ -3,12 +3,10 @@ import toast from "react-hot-toast";
 
 import styles from "@/app/_components/PortalComponents/EditSponsorCell/index.module.scss";
 import { trpc } from "@/trpc/react";
-import { type UserResource } from "@clerk/nextjs/types";
 
 import BasicButton, { ButtonVariant } from "../../Buttons/BasicButton";
 
 export interface DeleteOurWorkEntryProps {
-  currentUser: UserResource | undefined | null;
   currentRow: {
     id: number;
     year: number;
